+++ packages/medical-domain/src/domains/repositories/medical/datum.service.ts
@@ -46,7 +46,6 @@ import ZenModeService from './datum/zen-mode.service'
 import { isEpochBetweenBounds } from '../time/time.service'
 import { isBasal } from '../../models/medical/datum/basal.model'
 import { isDuration } from '../../models/medical/datum/basics/duration.model'
-import { isBg } from '../../models/medical/datum/bg.model'
 import { defaultWeekDaysFilter, type WeekDaysFilter } from '../../models/time/date-filter.model'
 import { DatumType } from '../../models/medical/datum/enums/datum-type.enum';
 import { DeviceEventSubtype } from '../../models/medical/datum/enums/device-event-subtype.enum';
@@ -110,14 +109,12 @@ const deduplicate = (data: Datum[], _opts: MedicalDataOptions): Datum[] => {
 export const filterOnDate = (data: Datum[], start: number, end: number, weekDaysFilter: WeekDaysFilter = defaultWeekDaysFilter): Datum[] => {
   return data.filter((dat: Datum) => {
     const epochStartInBounds = isEpochBetweenBounds(dat.epoch, start, end)
+    if (dat.isoWeekday !== undefined && !weekDaysFilter[dat.isoWeekday]) {
+      return false
+    }
     if (isBasal(dat) || isDuration(dat)) {
       const epochEndInBounds = isEpochBetweenBounds(dat.epochEnd, start, end)
       return epochStartInBounds || epochEndInBounds
-    }
-    if (isBg(dat)) {
-      if (!weekDaysFilter[dat.isoWeekday]) {
-        return false
-      }
     }
     return epochStartInBounds
   })